// src/app/client/model/user.ts
/**
 * my api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export class User {
    id?: string;
    email?: string;
    isEmailVerified?: boolean;
    name?: string;
    role?: User.RoleEnum;
}
export namespace User {
    export type RoleEnum = 0 | 1 | 2 | 3;
    export const RoleEnum = {
        Anon: 0 as RoleEnum,
        Regular: 1 as RoleEnum,
        Manager: 2 as RoleEnum,
        Admin: 3 as RoleEnum
    }
}